# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Dex Blog Newsletter API
  version: 1.0.0
  description: >-
    Public reader surface only. Editors manage blog posts and subscribers in Dex Web.
paths:
  /api/health:
    get:
      operationId: getHealth
      responses:
        '200':
          description: Application is healthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /api/application-info:
    get:
      operationId: getApplicationInfo
      responses:
        '200':
          description: Publication name and optional Dex Web link.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationInfo'
  /api/newsletter/subscriptions:
    post:
      operationId: subscribeToNewsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
      responses:
        '200':
          description: The address is subscribed, now or already.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/newsletter/unsubscriptions:
    post:
      operationId: unsubscribeFromNewsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnsubscriptionRequest'
      responses:
        '200':
          description: The address is no longer subscribed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnsubscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/Unavailable'
components:
  schemas:
    ApplicationInfo:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
        dexWebUrl:
          type: string
    SubscriptionRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          minLength: 1
          maxLength: 254
    SubscriptionResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [subscribed, already_subscribed]
    UnsubscriptionRequest:
      type: object
      additionalProperties: false
      required: [email, token]
      properties:
        email:
          type: string
          minLength: 1
          maxLength: 254
        token:
          type: string
          minLength: 1
          maxLength: 128
    UnsubscriptionResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [unsubscribed, not_subscribed]
    HealthResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        error:
          type: string
          pattern: '^[a-z][a-z0-9_]*$'
        message:
          type: string
          maxLength: 3000
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: The request conflicts with the subscriber list.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unavailable:
      description: The subscriber list is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'